+++ outputs.tf
@@ -2,12 +2,6 @@
 # OUTPUT CALCULATED VARIABLES (prefer full objects)
 # ------------------------------------------------------------------------------
 
-# remap iam to reduce one level of access (iam[]. instead of iam[].iam.)
-output "iam" {
-  description = "The iam resource objects that define the access to the resources."
-  value       = { for key, iam in module.iam : key => iam.iam }
-}
-
 # ------------------------------------------------------------------------------
 # OUTPUT ALL RESOURCES AS FULL OBJECTS
 # ------------------------------------------------------------------------------
@@ -15,17 +9,4 @@ output "iam" {
 output "repository" {
   description = "All `google_artifact_registry_repository` resource attributes."
   value       = try(google_artifact_registry_repository.repository[0], null)
-}
-
-# ------------------------------------------------------------------------------
-# OUTPUT ALL INPUT VARIABLES
-# ------------------------------------------------------------------------------
-
-# ------------------------------------------------------------------------------
-# OUTPUT MODULE CONFIGURATION
-# ------------------------------------------------------------------------------
-
-output "module_enabled" {
-  description = "Whether the module is enabled."
-  value       = var.module_enabled
 }