+++ variables.tf
@@ -50,20 +50,6 @@ variable "project" {
   default     = null
 }
 
-## IAM
-
-variable "iam" {
-  description = "(Optional) A list of IAM access."
-  type        = any
-  default     = []
-}
-
-variable "policy_bindings" {
-  description = "(Optional) A list of IAM policy bindings."
-  type        = any
-  default     = null
-}
-
 # ------------------------------------------------------------------------------
 # MODULE CONFIGURATION PARAMETERS
 # These variables are used to configure the module.